# docs/swagger.yaml
basePath: /api/v1
definitions:
  api.Result:
    properties:
      data:
        type: object
      error:
        $ref: '#/definitions/api.ResultError'
        type: object
      status:
        type: string
    type: object
  api.ResultError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  model.Pagination:
    properties:
      current:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  model.TableListData:
    properties:
      list:
        type: object
      pagination:
        $ref: '#/definitions/model.Pagination'
        type: object
    type: object
...
  service.ResourceActionEdit:
    properties:
      code:
        type: string
      description:
        type: string
      group:
        type: string
      name:
        type: string
      resource_server_id:
        type: integer
    type: object
  service.ResourceEditModel:
    properties:
      description:
        type: string
      name:
        type: string
      organization_id:
        type: integer
      secret:
        type: string
    type: object
  service.ResourceMenuEdit:
    properties:
      icon:
        type: string
      leaf:
        description: 是：子组件，否：是父组件
        type: integer
      name:
        type: string
      parent_id:
        type: integer
      resource_server_id:
        type: integer
      serial_number:
        type: integer
      url:
        type: string
    type: object
  service.ResourceRouteEdit:
    properties:
      method:
        type: string
      name:
        type: string
      path:
        type: string
      resource_server_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: 862860000@qq.com
    name: API Support
    url: https://github.com/nilorg/naas
  description: NilOrg认证授权服务Api详情.
  license:
    name: GNU General Public License v3.0
    url: https://github.com/nilorg/naas/blob/master/LICENSE
  termsOfService: https://github.com/nilorg
  title: NilOrg认证授权服务
  version: "1.0"
paths:
...
  /resource/actions:
    get:
      description: 查询资源服务器动作翻页数据
      parameters:
      - description: 当前页
        in: query
        name: current
        required: true
        type: integer
      - description: 页大小
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Result'
            - properties:
                data:
                  $ref: '#/definitions/model.TableListData'
              type: object
      security:
      - OAuth2AccessCode: []
      summary: 查询资源服务器动作
      tags:
      - ResourceAction（资源动作）
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.ResourceActionEdit'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Result'
      security:
      - OAuth2AccessCode: []
      summary: 添加动作
      tags:
      - ResourceAction（资源动作）
  /resource/actions/{resource_action_id}:
    delete:
      description: 根据资源动作ID,修改一个资源动作
      parameters:
      - description: resource action id
        in: path
        name: resource_action_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Result'
      security:
      - OAuth2AccessCode: []
      summary: 删除一个资源动作
      tags:
      - ResourceAction（资源动作）
    get:
      consumes:
      - application/json
      description: 根据菜单ID,获取一个动作
      parameters:
      - description: resource action id
        in: path
        name: resource_action_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Result'
      security:
      - OAuth2AccessCode: []
      summary: 获取一个动作
      tags:
      - ResourceAction（资源动作）
    put:
      description: 根据资源动作ID,修改一个资源动作
      parameters:
      - description: resource action id
        in: path
        name: resource_action_id
        required: true
        type: string
      - description: 资源动作需要修改的信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.ResourceActionEdit'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Result'
      security:
      - OAuth2AccessCode: []
      summary: 修改一个资源动作
      tags:
      - ResourceAction（资源动作）
  /resource/menus:
    get:
      consumes:
      - application/json
      description: 查询资源服务器菜单翻页数据
      parameters:
      - description: 当前页
        in: query
        name: current
        required: true
        type: integer
      - description: 页大小
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Result'
            - properties:
                data:
                  $ref: '#/definitions/model.TableListData'
              type: object
      security:
      - OAuth2AccessCode: []
      summary: 查询资源服务器菜单
      tags:
      - ResourceMenu（资源服务器菜单）
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.ResourceMenuEdit'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Result'
      security:
      - OAuth2AccessCode: []
      summary: 添加菜单
      tags:
      - ResourceRoute（资源服务器路由）
  /resource/routes:
    get:
      consumes:
      - application/json
      description: 查询资源服务器Rule翻页数据
      parameters:
      - description: 当前页
        in: query
        name: current
        required: true
        type: integer
      - description: 页大小
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Result'
            - properties:
                data:
                  $ref: '#/definitions/model.TableListData'
              type: object
      security:
      - OAuth2AccessCode: []
      summary: 查询资源服务器Rule
      tags:
      - ResourceRoute（资源服务器路由）
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.ResourceRouteEdit'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Result'
      security:
      - OAuth2AccessCode: []
      summary: 添加路由
      tags:
      - ResourceRoute（资源服务器路由）
  /resource/routes/{resource_menu_id}:
    delete:
      description: 根据资源菜单ID,修改一个资源菜单
      parameters:
      - description: resource menu id
        in: path
        name: resource_menu_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Result'
      security:
      - OAuth2AccessCode: []
      summary: 删除一个资源菜单
      tags:
      - ResourceMenu（资源服务器菜单）
    get:
      consumes:
      - application/json
      description: 根据菜单ID,获取一个菜单
      parameters:
      - description: resource menu id
        in: path
        name: resource_menu_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Result'
      security:
      - OAuth2AccessCode: []
      summary: 获取一个菜单
      tags:
      - ResourceMenu（资源服务器菜单）
    put:
      description: 根据资源菜单ID,修改一个资源菜单
      parameters:
      - description: resource menu id
        in: path
        name: resource_menu_id
        required: true
        type: string
      - description: 路由需要修改的信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.ResourceMenuEdit'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Result'
      security:
      - OAuth2AccessCode: []
      summary: 修改一个资源菜单
      tags:
      - ResourceMenu（资源服务器菜单）
  /resource/routes/{resource_route_id}:
    delete:
      consumes:
      - application/json
      description: 根据资源路由ID,删除一个资源路由
      parameters:
      - description: resource route id
        in: path
        name: resource_route_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Result'
      security:
      - OAuth2AccessCode: []
      summary: 删除一个资源路由
      tags:
      - ResourceRoute（资源服务器路由）
    get:
      consumes:
      - application/json
      description: 根据路由ID,获取一个路由
      parameters:
      - description: resource route id
        in: path
        name: resource_route_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Result'
      security:
      - OAuth2AccessCode: []
      summary: 获取一个路由
      tags:
      - ResourceRoute（资源服务器路由）
    put:
      consumes:
      - application/json
      description: 根据资源路由ID,修改一个资源路由
      parameters:
      - description: resource route id
        in: path
        name: resource_route_id
        required: true
        type: string
      - description: 路由需要修改的信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.ResourceRouteEdit'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Result'
      security:
      - OAuth2AccessCode: []
      summary: 修改一个资源路由
      tags:
      - ResourceRoute（资源服务器路由）
  /resource/servers:
    get:
      consumes:
      - application/json
      description: 查询资源翻页数据
      parameters:
      - description: 当前页
        in: query
        name: current
        required: true
        type: integer
      - description: 页大小
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Result'
            - properties:
                data:
                  $ref: '#/definitions/model.TableListData'
              type: object
      security:
      - OAuth2AccessCode: []
      summary: 查询资源
      tags:
      - Resource（资源）
    post:
      consumes:
      - application/json
      description: 创建资源
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.ResourceEditModel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Result'
      security:
      - OAuth2AccessCode: []
      summary: 创建资源
      tags:
      - Resource（资源）
  /resource/servers/{resource_server_id}:
    get:
      consumes:
      - application/json
      description: 根据资源ID,获取一个资源
      parameters:
      - description: resource id
        in: path
        name: resource_server_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Result'
      security:
      - OAuth2AccessCode: []
      summary: 获取一个资源
      tags:
      - Resource（资源）
    put:
      consumes:
      - application/json
      description: 根据资源ID,修改一个资源
      parameters:
      - description: resource id
        in: path
        name: resource_server_id
        required: true
        type: string
      - description: 资源需要修改的信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.ResourceEditModel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Result'
      security:
      - OAuth2AccessCode: []
      summary: 修改一个资源
      tags:
      - Resource（资源）
...
securityDefinitions:
  OAuth2AccessCode:
    authorizationUrl: http://localhost:8080/oauth2/authorize
    flow: accessCode
    scopes:
      email: ' 用户emial'
      openid: ' 用户openid'
      phone: ' 用户手机号'
      profile: ' 用户资料'
    tokenUrl: http://localhost:8080/oauth2/token
    type: oauth2
swagger: "2.0"